ipq_qca8075: support PHY LEDs on/off
Add code to manually manipulate the LEDs connected to the Ethernet PHY.

Current cTU hardware connects the LEDs to PHY port 0 on address 3. So
this code was tested with QCA8072_CTU_PHY_ADDR set to 3. This hardware
issue should be fixed in future hardware revision. So this commit sets
QCA8072_CTU_PHY_ADDR to 4 which is the address of the PHY active port 1.