qca: gpio: add support for generic gpio mux
Currently, there is only option to define a pin function for
serial, usb, etc. However, there is no way to actually do it
for regular GPIO, with no relation to any device.

This commit adds support for a generic, optional,
top level `/gpio` node that that device
will parse the list of gpios
and configure them accordingly.